requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
pydantic
python-decouple
python-jose
passlib[bcrypt]
python-multipart
reportlab
python-dotenv